+++ src/flow/create-single-project/single-create-project-flow.ts
@@ -1,11 +1,16 @@
+import * as vscode from 'vscode';
 import { Project } from "../../model/project/project";
 import { ProjectService } from "../../service/project-service";
+import { SelectProjectConfig, SelectProjectStep } from "../../step/project/select-project-step";
 import { StepBluePrint } from "../../step/step-blueprint";
 import { CreateProjectFlow } from "../create-project/create-project-flow";
 import { CreateProjectFlowConfig } from "../create-project/create-project-flow-config";
 
-export interface CreateSingleProjectFlowConfig extends CreateProjectFlowConfig<CreateSingleProjectFlowConfig>
+export interface CreateSingleProjectFlowConfig extends 
+	CreateProjectFlowConfig<CreateSingleProjectFlowConfig>,
+	SelectProjectConfig
 {
+	createdProject?: Project
 }
 
 export class CreateSingleProjectFlow extends CreateProjectFlow<CreateSingleProjectFlowConfig>
@@ -14,9 +19,10 @@ export class CreateSingleProjectFlow extends CreateProjectFlow<CreateSingleProje
 		return CreateSingleProjectFlow.continueProjectCreateFlow(_config, CreateSingleProjectFlow._onProjectCreationComplete);
 	}
 
-	private static _onProjectCreationComplete(_config: CreateSingleProjectFlowConfig, _project: Project): void
+	private static _onProjectCreationComplete(_config: CreateSingleProjectFlowConfig, _project: Project): undefined
 	{
 		const projectService = new ProjectService();
 		projectService.saveProject(_project);
+		return undefined;
 	}
 }