+++ src/model/command-info.ts
@@ -1,5 +1,0 @@
-export interface CommandInfo
-{
-	command: string;
-	title: string;
-}