+++ src/model/registry-project/registry-project.ts
@@ -1,8 +1,0 @@
-import * as vscode from 'vscode';
-import { Project } from '../project/project';
-
-export interface RegistryProject
-{
-	uri: vscode.Uri;
-	project: Project;
-}