+++ CHANGELOG.md
@@ -4,3 +4,9 @@
 
 - Initial release
 - Added documentation
+
+### v0.0.2 - June 23nd 2021
+
+- Added first release
+- Fixed directory issue with root projects
+- Removed 'platform' as required from schema as it isn't required for a rootproject